requirements.txt
pandas
numpy
scipy
plotly
langchain
langchain-groq
langchain-community
google-search-results
python-dotenv